HID: logitech-dj: add support for the Powerplay mat/receiver
The Logitech G Powerplay has a lightspeed receiver with a static HID++
device with ID 7 attached to it to. It is used to configure the led on
the mat. For this reason I increased the max number of devices.

I also marked all lightspeed devices as HID++ compatible. As the
internal powerplay device does not have REPORT_TYPE_KEYBOARD or
REPORT_TYPE_MOUSE it was not being marked as HID++ compatible in
logi_hidpp_dev_conn_notif_equad.

Signed-off-by: Filipe Laíns <lains@archlinux.org>